+++ src/index.ts
@@ -18,11 +18,12 @@ export * from './core/unwrapOr';
 export * from './core/unwrapOrElse';
 export * from './core/unwrapOrThrow';
 export * from './core/unwrapErr';
-export * from './core/expectResult';
+export * from './core/expectOk';
 export * from './core/expectErr';
 
 // Conversions
 export * from './core/fromPromise';
+export * from './core/fromSerialized';
 export * from './core/fromNullable';
 export * from './core/try';
 export * from './core/toPromise';